requirements.txt
boto3==1.26.145
diffusers==0.16.1
matplotlib==3.7.0
pillow==9.4.0
pyunsplash==1.0.0rc2
pyyaml==6.0
requests==2.28.1
safetensors==0.3.1
sagemaker==2.159.0
selenium==4.9.1
torch==1.12.1
webdriver-manager==3.8.6